# editoast: speed up layer generation