--- a/editoast/src/generated_data/sql/generate_operational_point_layer.sql
+++ b/editoast/src/generated_data/sql/generate_operational_point_layer.sql
@@ -9,35 +9,31 @@ WITH ops AS (
 ),
 collect AS (
     SELECT ops.op_id,
-        ST_Transform(
-            ST_LineInterpolatePoint(
-                ST_GeomFromGeoJSON(tracks.data->'geo'),
-                LEAST(
-                    GREATEST(
-                        ops.position / (tracks.data->'length')::float,
-                        0.
-                    ),
-                    1.
-                )
-            ),
-            3857
+        ST_LineInterpolatePoint(
+            tracks_layer.geographic,
+            LEAST(
+                GREATEST(
+                    ops.position / (tracks.data->'length')::float,
+                    0.
+                ),
+                1.
+            )
         ) AS geo,
-        ST_Transform(
-            ST_LineInterpolatePoint(
-                ST_GeomFromGeoJSON(tracks.data->'sch'),
-                LEAST(
-                    GREATEST(
-                        ops.position / (tracks.data->'length')::float,
-                        0.
-                    ),
-                    1.
-                )
-            ),
-            3857
+        ST_LineInterpolatePoint(
+            tracks_layer.schematic,
+            LEAST(
+                GREATEST(
+                    ops.position / (tracks.data->'length')::float,
+                    0.
+                ),
+                1.
+            )
         ) AS sch
     FROM ops
         INNER JOIN osrd_infra_tracksectionmodel AS tracks ON tracks.obj_id = ops.track_id
         AND tracks.infra_id = $1
+        INNER JOIN osrd_infra_tracksectionlayer AS tracks_layer ON tracks.obj_id = tracks_layer.obj_id
+        AND tracks.infra_id = tracks_layer.infra_id
 )
 INSERT INTO osrd_infra_operationalpointlayer (obj_id, infra_id, geographic, schematic)
 SELECT op_id,