--- a/editoast/src/generated_data/sql/insert_update_track_section_link_layer.sql
+++ b/editoast/src/generated_data/sql/insert_update_track_section_link_layer.sql
@@ -1,11 +1,13 @@
 WITH collect AS (
     SELECT links.obj_id AS link_id,
         links.data->'dst'->>'endpoint' AS ep,
-        ST_GeomFromGeoJSON(tracks.data->'geo') AS track_geo,
-        ST_GeomFromGeoJSON(tracks.data->'sch') AS track_sch
+        tracks_layer.geographic AS track_geo,
+        tracks_layer.schematic AS track_sch
     FROM osrd_infra_tracksectionlinkmodel AS links
         INNER JOIN osrd_infra_tracksectionmodel AS tracks ON tracks.obj_id = links.data->'dst'->>'track'
         AND tracks.infra_id = links.infra_id
+        INNER JOIN osrd_infra_tracksectionlayer AS tracks_layer ON tracks.obj_id = tracks_layer.obj_id
+        AND tracks.infra_id = tracks_layer.infra_id
     WHERE links.infra_id = $1
         AND links.obj_id = ANY($2)
 )
@@ -14,13 +16,13 @@ SELECT link_id,
     $1,
     CASE
         ep
-        WHEN 'BEGIN' THEN ST_Transform(ST_StartPoint(track_geo), 3857)
-        WHEN 'END' THEN ST_Transform(ST_EndPoint(track_geo), 3857)
+        WHEN 'BEGIN' THEN ST_StartPoint(track_geo)
+        WHEN 'END' THEN ST_EndPoint(track_geo)
     END,
     CASE
         ep
-        WHEN 'BEGIN' THEN ST_Transform(ST_StartPoint(track_sch), 3857)
-        WHEN 'END' THEN ST_Transform(ST_EndPoint(track_sch), 3857)
+        WHEN 'BEGIN' THEN ST_StartPoint(track_sch)
+        WHEN 'END' THEN ST_EndPoint(track_sch)
     END
 FROM collect ON CONFLICT (infra_id, obj_id) DO
 UPDATE