… editoast/src/generated_data/sql/insert_lpv_panel_layer.sql
… editoast/src/generated_data/sql/insert_operational_point_layer.sql
… editoast/src/generated_data/sql/insert_speed_section_layer.sql
… editoast/src/generated_data/sql/insert_update_buffer_stop_layer.sql
… editoast/src/generated_data/sql/insert_update_detector_layer.sql
… editoast/src/generated_data/sql/insert_update_switch_layer.sql