Uses track section layer avoiding using st_transform.

## Benchmark

|                  | Old Version   |   New version |
| ---------------- | ------------- | ------------- |
| Track section    |      3.99s    |       4.43s   |
| Speed section    |    160.21s    |      60.99s   |
| Signal           |      3.17s    |       3.43s   |
| Switch           |      2.13s    |       0.52s   |
| Buffer stop      |      0.35s    |       0.20s   |
| Catenary         |     10.02s    |       3.96s   |
| Detector         |     11.11s    |       3.17s   |
| Operational point|      4.53s    |       1.42s   |
| Track link       |      1.25s    |       0.29s   |
| Lpv panel        |      5.08s    |       2.21s   |
| Dead section     |      1.22s    |       0.63s   |
| **TOTAL**        |    **203s**   |     **81.2s** |